Added 'skeleton' for instructions and functions
In the parse_tokens() function there is now minimal support for each
built-in command. All 10 commands have still to be done, however. I have
also added the skeleton for the execute_process() function. This function
will be used to execute a new process in the shell, which is used when
calling external commands. A lot of TODO's have been added to the code!