Started the execute_process() code
The code for creating an external process has mostly been done - it has
yet to be fully tested, however. It follows the same logic as in figure
3.10 of the class book, except that it's "integrated" more into our shell.
execlp() is used because it takes into account the PATH variable and
accepts a list of arguments, which is perfect for our use case.