Fixed the external process execution
This means that stage two of the shell is now complete. I had to use execvp()
rather than execlp() because the latter does not accept a list of arguments
as a parameter. Also fixed a bug where if a command was not recognised it
would fail to terminate the child process.