Improved the alias code a bit
Also solved some compile errors. I'm sure some still exist but I'm calling
it a night. The same idea remains, it's best to use a map for the aliases.
However, I have run into some problems with pointers and memory allocation
that I'm not sure of how to resolve, so will read up on the best way to
approach it soon.