Implemented stage six
Stage six is now fully working. It uses two arrays: one which stores the
keys and another which stores the values (i.e. key->value). This maps
the aliases quite nicely to their commands. I've put everything into the
one file because it keeps it all in one place and it means we can just
submit 'main.c'. Perhaps we'll look into splitting it into several files
at the end, but this makes it easier for the time being. As a result, the
alias and unalias commands are fully implemented.